+++ lib/src/components/button.dart
@@ -7,6 +7,7 @@ import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/data.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
 import 'package:shadcn_ui/src/utils/debug_check.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 import 'package:shadcn_ui/src/utils/gesture_detector.dart';
 import 'package:shadcn_ui/src/utils/separated_iterable.dart';
 import 'package:shadcn_ui/src/utils/states_controller.dart';
@@ -76,6 +77,8 @@ class ShadButton extends StatefulWidget {
     this.textDirection,
     this.gap,
     this.onFocusChange,
+    this.orderPolicy,
+    this.expands,
   }) : variant = ShadButtonVariant.primary;
 
   const ShadButton.raw({
@@ -128,6 +131,8 @@ class ShadButton extends StatefulWidget {
     this.textDirection,
     this.gap,
     this.onFocusChange,
+    this.orderPolicy,
+    this.expands,
   });
 
   const ShadButton.destructive({
@@ -179,6 +184,8 @@ class ShadButton extends StatefulWidget {
     this.textDirection,
     this.gap,
     this.onFocusChange,
+    this.orderPolicy,
+    this.expands,
   }) : variant = ShadButtonVariant.destructive;
 
   const ShadButton.outline({
@@ -230,6 +237,8 @@ class ShadButton extends StatefulWidget {
     this.textDirection,
     this.gap,
     this.onFocusChange,
+    this.orderPolicy,
+    this.expands,
   }) : variant = ShadButtonVariant.outline;
 
   const ShadButton.secondary({
@@ -281,6 +290,8 @@ class ShadButton extends StatefulWidget {
     this.textDirection,
     this.gap,
     this.onFocusChange,
+    this.orderPolicy,
+    this.expands,
   }) : variant = ShadButtonVariant.secondary;
 
   const ShadButton.ghost({
@@ -332,6 +343,8 @@ class ShadButton extends StatefulWidget {
     this.textDirection,
     this.gap,
     this.onFocusChange,
+    this.orderPolicy,
+    this.expands,
   }) : variant = ShadButtonVariant.ghost;
 
   const ShadButton.link({
@@ -382,6 +395,8 @@ class ShadButton extends StatefulWidget {
     this.textDirection,
     this.gap,
     this.onFocusChange,
+    this.orderPolicy,
+    this.expands,
   })  : variant = ShadButtonVariant.link,
         icon = null;
 
@@ -411,6 +426,12 @@ class ShadButton extends StatefulWidget {
   final ShadDecoration? decoration;
   final bool enabled;
   final ShadStatesController? statesController;
+
+  /// {@template ShadButton.orderPolicy}
+  /// The order policy of the items that compose the button, defaults to
+  /// [WidgetOrderPolicy.linear()].
+  /// {@endtemplate}
+  final WidgetOrderPolicy? orderPolicy;
 
   /// {@template ShadButton.gap}
   /// The gap between the icon and the text.
@@ -452,6 +473,11 @@ class ShadButton extends StatefulWidget {
   final Duration? longPressDuration;
   final TextDirection? textDirection;
   final ValueChanged<bool>? onFocusChange;
+
+  /// {@template ShadButton.expands}
+  /// Whether the child expands to fill the available space, defaults to false.
+  /// {@endtemplate}
+  final bool? expands;
 
   @override
   State<ShadButton> createState() => _ShadButtonState();
@@ -708,6 +734,13 @@ class _ShadButtonState extends State<ShadButton> {
 
     final effectiveGap = widget.gap ?? buttonTheme(theme).gap ?? 8;
 
+    final effectiveOrderPolicy = widget.orderPolicy ??
+        buttonTheme(theme).orderPolicy ??
+        const WidgetOrderPolicy.linear();
+
+    final effectiveExpands =
+        widget.expands ?? buttonTheme(theme).expands ?? false;
+
     return CallbackShortcuts(
       bindings: {
         const SingleActivator(LogicalKeyboardKey.enter): onTap,
@@ -744,6 +777,31 @@ class _ShadButtonState extends State<ShadButton> {
               child: icon,
             );
           }
+
+          Widget? child = widget.child == null
+              ? null
+              : DefaultTextStyle(
+                  style: theme.textTheme.small.copyWith(
+                    color: hasPressedForegroundColor && pressed
+                        ? pressedForegroundColor(theme)
+                        : hovered
+                            ? hoverForeground(theme)
+                            : foreground(theme),
+                    decoration: textDecoration(
+                      theme,
+                      hovered: hovered,
+                    ),
+                    decorationColor: foreground(theme),
+                    decorationStyle: TextDecorationStyle.solid,
+                  ),
+                  textAlign: TextAlign.center,
+                  child: widget.child!,
+                );
+
+          if (child != null && effectiveExpands) {
+            child = Expanded(child: child);
+          }
+
           return Semantics(
             container: true,
             button: true,
@@ -815,25 +873,10 @@ class _ShadButtonState extends State<ShadButton> {
                           textDirection: effectiveTextDirection,
                           children: [
                             if (icon != null) icon,
-                            if (widget.child != null)
-                              DefaultTextStyle(
-                                style: theme.textTheme.small.copyWith(
-                                  color: hasPressedForegroundColor && pressed
-                                      ? pressedForegroundColor(theme)
-                                      : hovered
-                                          ? hoverForeground(theme)
-                                          : foreground(theme),
-                                  decoration: textDecoration(
-                                    theme,
-                                    hovered: hovered,
-                                  ),
-                                  decorationColor: foreground(theme),
-                                  decorationStyle: TextDecorationStyle.solid,
-                                ),
-                                textAlign: TextAlign.center,
-                                child: widget.child!,
-                              ),
-                          ].separatedBy(SizedBox(width: effectiveGap)),
+                            if (child != null) child,
+                          ]
+                              .order(effectiveOrderPolicy)
+                              .separatedBy(SizedBox(width: effectiveGap)),
                         ),
                       ),
                     ),