+++ lib/src/components/checkbox.dart
@@ -8,6 +8,7 @@ import 'package:shadcn_ui/src/raw_components/focusable.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
 import 'package:shadcn_ui/src/utils/debug_check.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 
 class ShadCheckbox extends StatefulWidget {
   const ShadCheckbox({
@@ -26,6 +27,7 @@ class ShadCheckbox extends StatefulWidget {
     this.padding,
     this.direction,
     this.crossAxisAlignment,
+    this.orderPolicy,
   });
 
   /// Whether the checkbox is on or off.
@@ -75,6 +77,12 @@ class ShadCheckbox extends StatefulWidget {
   /// both not null, and [CrossAxisAlignment.center] otherwise.
   /// {@endtemplate}
   final CrossAxisAlignment? crossAxisAlignment;
+
+  /// {@template ShadCheckbox.orderPolicy}
+  /// The order policy of the items that compose the checkbox, defaults to
+  /// [WidgetOrderPolicy.linear()].
+  /// {@endtemplate}
+  final WidgetOrderPolicy? orderPolicy;
 
   @override
   State<ShadCheckbox> createState() => _ShadCheckboxState();
@@ -133,6 +141,10 @@ class _ShadCheckboxState extends State<ShadCheckbox> {
     final effectivePadding = widget.padding ??
         theme.checkboxTheme.padding ??
         const EdgeInsets.only(left: 8);
+
+    final effectiveOrderPolicy = widget.orderPolicy ??
+        theme.checkboxTheme.orderPolicy ??
+        const WidgetOrderPolicy.linear();
 
     final checkbox = Semantics(
       checked: widget.value,
@@ -225,7 +237,7 @@ class _ShadCheckboxState extends State<ShadCheckbox> {
                   ),
                 ),
               ),
-          ],
+          ].order(effectiveOrderPolicy),
         ),
       ),
     );