+++ lib/src/components/form/fields/checkbox.dart
@@ -3,6 +3,7 @@ import 'package:shadcn_ui/src/components/checkbox.dart';
 import 'package:shadcn_ui/src/components/form/field.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 
 class ShadCheckboxFormField extends ShadFormBuilderField<bool> {
   ShadCheckboxFormField({
@@ -33,6 +34,9 @@ class ShadCheckboxFormField extends ShadFormBuilderField<bool> {
 
     /// {@macro ShadCheckbox.crossAxisAlignment}
     CrossAxisAlignment? crossAxisAlignment,
+
+    /// {@macro ShadCheckbox.orderPolicy}
+    WidgetOrderPolicy? orderPolicy,
   }) : super(
           initialValue: initialValue,
           onChanged: onChanged == null ? null : (v) => onChanged(v ?? false),
@@ -58,6 +62,7 @@ class ShadCheckboxFormField extends ShadFormBuilderField<bool> {
               direction: direction,
               decoration: state.decoration,
               crossAxisAlignment: crossAxisAlignment,
+              orderPolicy: orderPolicy,
             );
           },
         );